add utils.js file
