+++ package.json
@@ -29,10 +29,11 @@
     "webpack-dev-server": "^3.9.0"
   },
   "scripts": {
-    "start": "nodemon server/tempserver.js",
+    "start": "nodemon server/server.js",
+    "stats": "webpack --profile --json > compilation-stats.json",
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "NODE_ENV=production webpack",
-    "dev": "concurrently \"cross-env NODE_ENV=development webpack-dev-server --open\" \"cross-env nodemon server/tempserver.js || true\""
+    "dev": "concurrently \"cross-env NODE_ENV=development webpack-dev-server --open\" \"cross-env nodemon server/server.js || true\""
   },
   "keywords": [],
   "author": "CoNST",
